requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
langchain-neo4j>=0.0.1
python-dotenv>=1.0.0
neo4j>=5.15.0
google-generativeai>=0.3.2
openai>=1.0.0
flask>=2.0.0 